+++ frontend_httpc/src/pages/messageDisplay.tsx
@@ -1,11 +1,11 @@
 import { useState, useEffect } from "react";
-import { useWebSocket } from "../services/network.tsx";
+import { useWebSocket } from "../services/handleNetwork.tsx";
 import NetworkStatus from "../components/networkStatus.tsx";
 import MessageBox from "../components/messageBox.tsx";
 import MessageFeed from "../components/messageFeed.tsx";
 import UserSelection from "../components/userSelection.tsx";
-import { generateId } from "../services/generateRandom";
-import { setInLocalStorage } from "../services/browserStorage";
+import { generateId } from "../services/handleUser.tsx";
+import { setInLocalStorage } from "../services/handleBrowser.tsx";
 import "./messageDisplay.css";
 
 
@@ -32,8 +32,12 @@ function MessageDisplay() {
       username = generateId(5);
       setInLocalStorage("username", username);
     }
+    // TODO: make initial register post request in this use effect
+
+
     setUserName(username || "");
   }, [userName]);
+
 
   return (
     <>
