+++ frontend_httpc/src/services/handleMessages.tsx
@@ -1,3 +1,6 @@
+import axios from "axios"; 
+
+
 type SocketRef = React.RefObject<WebSocket | null>;
 type MessageObject = {
   sender_id?: string;
@@ -69,5 +72,28 @@ export const sendMessage = async ({
     }
   } else {
     console.log("Cannot send message. WebSocket is not open.");
+  }
+};
+
+export const getAllMessages = async () => {
+  try {
+    const response = await axios.get("http://127.0.0.1:8081/messages");
+    const msgData = response.data;
+    console.log("message data", msgData);
+
+    let msgString = "";
+    for (let i = 0; i < msgData.length; i++) {
+      const c = msgData[i];
+      if (msgData[i - 1] === " " || msgData[i - 1] === ":" || msgData[i - 1] === "{" || msgData[i + 1] === "}" || msgData[i + 1] === ":" || msgData[i + 1] === ","){
+        msgString += c === "'" ? '"' : c;
+      } else {
+        msgString += c;
+      }
+    }
+    console.log("reconstructed string:", msgString);
+    return JSON.parse(msgString);
+
+  } catch (error) {
+    console.log("Error getting response:", error);
   }
 };
