+++ frontend_httpc/src/services/handleNetwork.tsx
